[feat] Add CSA attention PyPTO kernel
## summary

- Add CSA attention prefill/decode kernels, golden validation,
  runner specs, and official-model tests.
- Reuse shared window-cache and compressed KV-pool helpers across
  SWA, HCA, and CSA attention paths.
- Adjust ratio4 compressor/indexer behavior and broaden SWA/HCA/CSA
  golden test coverage for prefill and decode cases.

+++ models/indexer.py
@@ -533,7 +533,7 @@ def golden_indexer_forward(tensors, start_pos: int):
     golden_compressor_ratio4_indexer_forward(comp_tensors, start_pos=start_pos)
 
     weights = torch.matmul(tensors["x"].float(), tensors["weights_proj_t"].float()).to(torch.bfloat16)
-    weights = weights.float() * INDEX_WEIGHTS_SCALE
+    weights = (weights * INDEX_WEIGHTS_SCALE).to(torch.bfloat16).float()
 
     score_full = torch.full((B, seq_len, INDEX_SCORE_LEN), NEG_INF, dtype=torch.float32)
     topk = torch.full((B, seq_len, INDEX_TOPK), -1, dtype=torch.int32)
